Major commit. Static part of gui should now be almost completely implemented
Broad strokes changes
Completely implemented the static parts of the gui as might be expected. Two slots in the top one for model and one for pristine matter.
Added the player inventory which will automatically center below the main gui.

changes in GuiDigitalAgonizer
  -Renamed TEXTURE to something more descriptive
  -Refactored HEIGHT -> TOTAL_HEIGHT
   and WIDTH -> TOTAL WIDTH to be more descriptive
   also changed their values
  -Implemented player inventory in the gui
  Very important that the formula be what it is so it stays centered. Im not sure why there isnt a version of drawPlayerInventory that does this automatically.
Added GuiTextures.java as of now im just adding it to force me to make a choice about weather to move all the gui textures to a single place or not.
Changes in TileEntityDigitalAgonizer.java
-added item handlers for things
-Pristine matter item handler should be temporary should change to something that handles any valid matter type for this block (probably).
-Copied over the capability adapter from LootFabricator, I dont know exactly how this works I do wonder if it isnt possible to make it more modular.